1ファイル1コンポーネントで修正（まだテストしてない）
